Refactor: Unify session events and blacklist endpoints
- Consolidate `session_start` and `session_end` MMP events into a single `user_session` event. This new event includes the session duration.
- Blacklist the old `/session_start` and `/session_end` endpoints to prevent them from being retried if they fail.
- Add the app's version code to MMP events and other backend requests.
- Cache the app's version code in `WalletUtils` for efficiency.
- Change the return type of `GetAppInstalledVersion` from `Int` to `Long` to correctly handle `longVersionCode`.

+++ android-appcoins-billing/src/main/java/com/appcoins/sdk/billing/repositories/StoreLinkMapperRepository.kt
@@ -105,7 +105,7 @@ class StoreLinkMapperRepository(private val bdsService: BdsService) {
         packageName: String,
         appInstallerPackageName: String?,
         oemid: String?,
-        versionCode: Int,
+        versionCode: Long,
         q: String?,
     ): NewVersionAvailableResponse {
         val countDownLatch = CountDownLatch(1)
